refactor(web): split live update modules [OPE-321]
